Modify the way physical and fluid domains are initialized in solver
Changelog:

- Add function InitPhysical in solver.{cpp/hpp}
- InitPhysical and InitFluid now handle the task of initializing the domains
- Modify Init function in Solver so that it only initializes the solver modules
- Expose InitPhysical function in wrapper.cpp
- Update ParametricSolver with the new initialization format

This change is a precursor to adding pickle storage/restoration for the Solver
class. The new Init function can be called in order to initialize the solver
modules without touching the domains. This allows for the domains to be restored
separately from pickle storage, followed by a call to Init to re-initialize the
solver modules. Since all persistent information about the state of the solver
is stored in the domain and common parameter objects, this allows for a smooth
restart of the solver. This new format has been checked for a simple case, but
not for the case of multiple replicates involving spatially varying boundaries.