Implement caching and restarting of dynamic analysis solver
Changelog:

- Modify psolver.py to incorporate caching of results. This involves adding a
  new function _run_solver, where the cache is read/written. The solver is
  cached in a pickle file identified by the replicate index and store in a local
  folder called "cache". Cache contains the result history for all previous time
  steps and the solver state at the current time step.
- The constant CACHE_SAVE_FREQUENCY in psolver.py controls the frequency at
  which caching is done
- The use_cache option to ParametricSolverMaster enables caching
- Modify wrapper.cpp to be consistent with the change of pointers to smart
  pointers. The property definition lines that expose the FEM domains as
  arguments of Solver have been simplified following the change to smart pointers

Dynamic analysis now works with caching of results at intermediate time steps.
Solver restarts have been tested for a few cases and the results show no
significant deviation from the continuous (non-restarted) solution. A backup
procedure has been incorporated so that a backup of the cache files is created
before writing to them, to ensure that data corruption does not occur. Restarts
with multiple replicates in a multiprocessing setup needs to be tested.