Gracefully handle errors during solver restore process
Changelog:

- Check for errors during file restore process is ParametricSolver. If an error
  occurs, then ignore the restore process and remove the backup file.

This change allows the solver to continue despite errors in the backup file that
may have occured during a previous solver run. If the solver is unable to
restore the state from the backup file, it is ignored and the file is deleted so
that the solver can start afresh on that replicate.