Try new algorithm to detect pull-in during dynamic analysis
Changelog:

- Introduce a counter variable and compare err with prevErr only if counter is
  greater than 2

The number of cases where the displacement actually exceed the gap is small and
so this alone is not a good test for pull-in. By using a counter variable, the
test is relient to false positives where the error rises during the first 1 or 2
steps and then dies down.